partial split of of CLIP code and pre-run code. Updated requrements file.
