+++ requirements.txt
@@ -1,10 +1,12 @@
 # local package
 -e .
 
-# external requirements
-click
-Sphinx
-coverage
-awscli
-flake8
-python-dotenv>=0.5.1
+click==8.0.4
+datasets==2.7.1
+Markdown==3.4.1
+Pillow==9.4.0
+python-dotenv==0.21.0
+setuptools==65.5.0
+torch==1.13.1
+torchvision==0.14.1
+transformers==4.25.1
